+++ GameFramework/Src/Mesh.cpp
@@ -7,7 +7,7 @@ MeshProxy* ColoredMesh::CreateMeshProxy()
 {
 	MeshProxy* meshProxy = new MeshProxy();
 
-	meshProxy->Initialize(&Vertices[0], Vertices.size(),  sizeof(ColoredVertex),  &Indices[0], Indices.size());
+	meshProxy->Initialize(&Vertices[0], Vertices.size(),  static_cast<UINT>(sizeof(ColoredVertex)),  &Indices[0], Indices.size());
 
 	return meshProxy;
 }
@@ -16,7 +16,7 @@ MeshProxy* TexturedMesh::CreateMeshProxy()
 {
 	MeshProxy* meshProxy = new MeshProxy();
 
-	meshProxy->Initialize(&Vertices[0], Vertices.size(), sizeof(TexturedVertex), &Indices[0], Indices.size());
+	meshProxy->Initialize(&Vertices[0], Vertices.size(), static_cast<UINT>(sizeof(TexturedVertex)), &Indices[0], Indices.size());
 
 	return meshProxy;
 }
@@ -25,7 +25,7 @@ MeshProxy* BasicMesh::CreateMeshProxy()
 {
 	MeshProxy* meshProxy = new MeshProxy();
 
-	meshProxy->Initialize(&Vertices[0], Vertices.size(), sizeof(BasicVertex), &Indices[0], Indices.size());
+	meshProxy->Initialize(&Vertices[0], Vertices.size(), static_cast<UINT>(sizeof(BasicVertex)), &Indices[0], Indices.size());
 
 	return meshProxy;
 }
